Load to DB Products

+++ pages/api/seed.js
@@ -5,13 +5,17 @@ import Product from "../../models/Product";
 
 /*Загрузка данных в БД*/
 const handler = async (req, res) => {
-    await db.connect();
-    await User.deleteMany();
-    await User.insertMany(data.users);
-    await Product.deleteMany();
-    await Product.insertMany(data.products);
-    await db.disconnect();
-    res.send({message: "Данные загружены"});
+    if (req.query.password === process.env.STUPID_PASSWORD) {
+        await db.connect();
+        await User.deleteMany();
+        await User.insertMany(data.users);
+        await Product.deleteMany();
+        await Product.insertMany(data.products);
+        await db.disconnect();
+        res.send({message: "Данные загружены."});
+    } else {
+        res.send({message: "А ты неплох. В доступе отказано."})
+    }
 };
 
 export default handler
